DEV-2766 fix linter
