+++ cmd/rport/servicemanagement/service.go
@@ -1,4 +1,4 @@
-package service_management
+package servicemanagement
 
 import (
 	"context"
@@ -7,21 +7,17 @@ import (
 
 	"github.com/kardianos/service"
 
-	chserver "github.com/cloudradar-monitoring/rport/server"
+	chclient "github.com/cloudradar-monitoring/rport/client"
 	chshare "github.com/cloudradar-monitoring/rport/share"
 )
 
 var svcConfig = &service.Config{
-	Name:        "rportd",
-	DisplayName: "Rport Server",
+	Name:        "rport",
+	DisplayName: "Rport Client",
 	Description: "Create reverse tunnels with ease.",
-	Option: service.KeyValue{
-		"LimitNOFILE":         1048576,
-		"AmbientCapabilities": "CAP_NET_BIND_SERVICE",
-	},
 }
 
-func HandleSvcCommand(svcCommand string, configPath string, user *string) error {
+func HandleSvcCommand(svcCommand string, configPath string, user string) error {
 	svc, err := getService(nil, configPath, user)
 	if err != nil {
 		return err
@@ -30,8 +26,8 @@ func HandleSvcCommand(svcCommand string, configPath string, user *string) error 
 	return chshare.HandleServiceCommand(svc, svcCommand)
 }
 
-func RunAsService(s *chserver.Server, configPath string) error {
-	svc, err := getService(s, configPath, nil)
+func RunAsService(c *chclient.Client, configPath string) error {
+	svc, err := getService(c, configPath, "")
 	if err != nil {
 		return err
 	}
@@ -39,29 +35,31 @@ func RunAsService(s *chserver.Server, configPath string) error {
 	return svc.Run()
 }
 
-func getService(s *chserver.Server, configPath string, user *string) (service.Service, error) {
+func getService(c *chclient.Client, configPath string, user string) (service.Service, error) {
 	absConfigPath, err := filepath.Abs(configPath)
 	if err != nil {
 		return nil, err
 	}
 	svcConfig.Arguments = []string{"-c", absConfigPath}
-	if user != nil {
-		svcConfig.UserName = *user
+	if user != "" {
+		svcConfig.UserName = user
 	}
-	return service.New(&serviceWrapper{s}, svcConfig)
+	return service.New(&serviceWrapper{c}, svcConfig)
 }
 
 type serviceWrapper struct {
-	*chserver.Server
+	*chclient.Client
 }
 
 func (w *serviceWrapper) Start(service.Service) error {
-	if w.Server == nil {
+	if w.Client == nil {
 		return nil
 	}
 	go func() {
-		ctx := context.Background()
-		if err := w.Server.Run(ctx); err != nil {
+		ctx, cancel := context.WithCancel(context.Background())
+		defer cancel()
+
+		if err := w.Client.Run(ctx); err != nil {
 			log.Println(err)
 		}
 	}()
@@ -69,5 +67,5 @@ func (w *serviceWrapper) Start(service.Service) error {
 }
 
 func (w *serviceWrapper) Stop(service.Service) error {
-	return w.Server.Close()
+	return w.Client.Close()
 }
